Use byte count in Ym laser render

+++ src/pppYmLaser.cpp
@@ -119,7 +119,7 @@ extern "C" void pppRenderYmLaser(pppYmLaser* laser, pppYmLaserUnkB* step, _pppCt
 	int colorOffset = serializedDataOffsets[1];
 	pppYmLaserColorData* colorData = (pppYmLaserColorData*)((u8*)laser + 0x80 + colorOffset);
 	s32 dataValIndex = step->m_dataValIndex;
-	int count;
+	u8 count;
 	s32 i;
 	s32 alphaStep;
 	char alphaMax;
